move enum to jwt package
